main_service/requirements.txt
requests
yt-dlp
faster-whisper
instaloader
moviepy==1.0.3
openai
instagrapi 
tiktok-uploader 
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
python-telegram-bot
python-dotenv
openai-whisper